+++ Changelog.md
@@ -20,3 +20,16 @@
 * Config Class
     * rename list_virtual_centers to get_virtual_centers
     * Made the update_settings_general work properly
+
+10-02-2021
+* Minor bug fixes
+* Inventory Class
+    * renamed list_ad_domains to get_ad_domains
+* Config Class
+    * Added update_settings
+    * Added update_settings_security
+    * Added update_settings_general
+* External Class
+    * Renamed list_hvpools to get_hvpools
+    * Started work on get_base_vms
+        * work in progress!!